Moves folders for incident priority and type backend components inside incident
