+++ src/dispatch/incident/service.py
@@ -14,10 +14,10 @@ from dispatch.case import service as case_service
 from dispatch.database.core import SessionLocal
 from dispatch.event import service as event_service
 from dispatch.exceptions import NotFoundError
+from dispatch.incident.priority import service as incident_priority_service
+from dispatch.incident.type import service as incident_type_service
 from dispatch.incident_cost import service as incident_cost_service
-from dispatch.incident_priority import service as incident_priority_service
 from dispatch.incident_role.service import resolve_role
-from dispatch.incident_type import service as incident_type_service
 from dispatch.participant import flows as participant_flows
 from dispatch.participant_role.models import ParticipantRoleType
 from dispatch.plugin import service as plugin_service
