+++ tests/incident_priority/test_incident_priority_service.py
@@ -1,12 +1,12 @@
 def test_get(session, incident_priority):
-    from dispatch.incident_priority.service import get
+    from dispatch.incident.priority.service import get
 
     t_incident_priority = get(db_session=session, incident_priority_id=incident_priority.id)
     assert t_incident_priority.id == incident_priority.id
 
 
 def test_get_by_name(session, incident_priority):
-    from dispatch.incident_priority.service import get_by_name
+    from dispatch.incident.priority.service import get_by_name
 
     t_incident_priority = get_by_name(
         db_session=session, project_id=incident_priority.project.id, name=incident_priority.name
@@ -15,7 +15,7 @@ def test_get_by_name(session, incident_priority):
 
 
 def test_get_all(session, project, incident_priorities):
-    from dispatch.incident_priority.service import get_all
+    from dispatch.incident.priority.service import get_all
 
     t_incident_priorities = get_all(
         db_session=session, project_id=incident_priorities[0].project.id
@@ -24,8 +24,8 @@ def test_get_all(session, project, incident_priorities):
 
 
 def test_create(session, project):
-    from dispatch.incident_priority.service import create
-    from dispatch.incident_priority.models import IncidentPriorityCreate
+    from dispatch.incident.priority.service import create
+    from dispatch.incident.priority.models import IncidentPriorityCreate
 
     name = "XXX"
     description = "XXXXXX"
@@ -40,8 +40,8 @@ def test_create(session, project):
 
 
 def test_update(session, incident_priority):
-    from dispatch.incident_priority.service import update
-    from dispatch.incident_priority.models import IncidentPriorityUpdate
+    from dispatch.incident.priority.service import update
+    from dispatch.incident.priority.models import IncidentPriorityUpdate
 
     name = "Updated incident priority name"
 
@@ -55,7 +55,7 @@ def test_update(session, incident_priority):
 
 
 def test_delete(session, incident_priority):
-    from dispatch.incident_priority.service import delete, get
+    from dispatch.incident.priority.service import delete, get
 
     delete(db_session=session, incident_priority_id=incident_priority.id)
     assert not get(db_session=session, incident_priority_id=incident_priority.id)
