+++ tests/incident_type/test_incident_type_service.py
@@ -1,12 +1,12 @@
 def test_get(session, incident_type):
-    from dispatch.incident_type.service import get
+    from dispatch.incident.type.service import get
 
     t_incident_type = get(db_session=session, incident_type_id=incident_type.id)
     assert t_incident_type.id == incident_type.id
 
 
 def test_get_by_name(session, incident_type):
-    from dispatch.incident_type.service import get_by_name
+    from dispatch.incident.type.service import get_by_name
 
     t_incident_type = get_by_name(
         db_session=session, project_id=incident_type.project.id, name=incident_type.name
@@ -15,7 +15,7 @@ def test_get_by_name(session, incident_type):
 
 
 def test_get_by_slug(session, incident_type):
-    from dispatch.incident_type.service import get_by_slug
+    from dispatch.incident.type.service import get_by_slug
 
     t_incident_type = get_by_slug(
         db_session=session, project_id=incident_type.project.id, slug=incident_type.slug
@@ -24,15 +24,15 @@ def test_get_by_slug(session, incident_type):
 
 
 def test_get_all(session, project, incident_types):
-    from dispatch.incident_type.service import get_all
+    from dispatch.incident.type.service import get_all
 
     t_incident_types = get_all(db_session=session, project_id=incident_types[0].project.id).all()
     assert t_incident_types
 
 
 def test_create(session, project, document):
-    from dispatch.incident_type.service import create
-    from dispatch.incident_type.models import IncidentTypeCreate
+    from dispatch.incident.type.service import create
+    from dispatch.incident.type.models import IncidentTypeCreate
 
     name = "XXX"
 
@@ -47,8 +47,8 @@ def test_create(session, project, document):
 
 
 def test_update(session, incident_type):
-    from dispatch.incident_type.service import update
-    from dispatch.incident_type.models import IncidentTypeUpdate
+    from dispatch.incident.type.service import update
+    from dispatch.incident.type.models import IncidentTypeUpdate
 
     name = "Updated incident type name"
 
@@ -62,7 +62,7 @@ def test_update(session, incident_type):
 
 
 def test_delete(session, incident_type):
-    from dispatch.incident_type.service import delete, get
+    from dispatch.incident.type.service import delete, get
 
     delete(db_session=session, incident_type_id=incident_type.id)
     assert not get(db_session=session, incident_type_id=incident_type.id)
